Feat : window size hook 생성, 사용용
