+++ src/components/Layout/Sidebar/Sidebar.jsx
@@ -1,17 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import useWindow from '../../../hooks/useWindow';
 import SidebarLarge from './SidebarLarge';
 import SidebarSmall from './SidebarSmall';
 
 export default function Sidebar({ sidebar }) {
-  const [resize, setResize] = useState(window.innerWidth);
-  useEffect(() => {
-    const listener = () => {
-      setResize(window.innerWidth);
-    };
-
-    window.addEventListener('resize', listener);
-    return () => window.removeEventListener('resize', listener);
-  }, []);
+  const resize = useWindow();
   return (
     <>{resize >= 1300 ? sidebar ? <SidebarSmall /> : <SidebarLarge /> : resize >= 792 ? <SidebarSmall /> : null}</>
   );
